+++ src/OmniSharp.Script/ScriptHelper.cs
@@ -16,6 +16,7 @@ namespace OmniSharp.Script
     {
         private const string BinderFlagsType = "Microsoft.CodeAnalysis.CSharp.BinderFlags";
         private const string TopLevelBinderFlagsProperty = "TopLevelBinderFlags";
+        private const string ReferencesSupersedeLowerVersionsProperty = "ReferencesSupersedeLowerVersions_internal_protected_set";
         private const string IgnoreCorLibraryDuplicatedTypesField = "IgnoreCorLibraryDuplicatedTypes";
         private const string RuntimeMetadataReferenceResolverType = "Microsoft.CodeAnalysis.Scripting.Hosting.RuntimeMetadataReferenceResolver";
         private const string ResolverField = "_resolver";
@@ -73,6 +74,11 @@ namespace OmniSharp.Script
                 topLevelBinderFlagsProperty?.SetValue(compilationOptions, ignoreCorLibraryDuplicatedTypesValue);
             }
 
+            // in scripts, the option to supersede lower versions is ALWAYS enabled
+            // see: https://github.com/dotnet/roslyn/blob/version-2.6.0-beta3/src/Compilers/CSharp/Portable/Compilation/CSharpCompilation.cs#L199
+            var referencesSupersedeLowerVersionsProperty = typeof(CompilationOptions).GetProperty(ReferencesSupersedeLowerVersionsProperty, BindingFlags.Instance | BindingFlags.NonPublic);
+            referencesSupersedeLowerVersionsProperty?.SetValue(compilationOptions, true);
+
             return compilationOptions;
         }
 
@@ -93,9 +99,10 @@ namespace OmniSharp.Script
                 : new CachingScriptMetadataResolver(_resolver);
         }
  
-        public ProjectInfo CreateProject(string csxFileName, IEnumerable<MetadataReference> references, IEnumerable<string> namespaces = null)
+        public ProjectInfo CreateProject(string csxFileName, IEnumerable<MetadataReference> references, string csxFilePath, IEnumerable<string> namespaces = null)
         {
             var project = ProjectInfo.Create(
+                filePath: csxFilePath,
                 id: ProjectId.CreateNewId(),
                 version: VersionStamp.Create(),
                 name: csxFileName,