+++ tests/OmniSharp.Roslyn.CSharp.Tests/FormattingFacts.cs
@@ -18,8 +18,8 @@ namespace OmniSharp.Roslyn.CSharp.Tests
 {
     public class FormattingFacts : AbstractTestFixture
     {
-        public FormattingFacts(ITestOutputHelper output)
-            : base(output)
+        public FormattingFacts(ITestOutputHelper output, SharedOmniSharpHostFixture sharedOmniSharpHostFixture)
+            : base(output, sharedOmniSharpHostFixture)
         {
         }
 
@@ -268,37 +268,34 @@ class C {
         private async Task AssertTextChanges(string source, params LinePositionSpanTextChange[] expected)
         {
             var testFile = new TestFile("dummy.cs", source);
+            SharedOmniSharpTestHost.AddFilesToWorkspace(testFile);
+            var span = testFile.Content.GetSpans().Single();
+            var range = testFile.Content.GetRangeFromSpan(span);
 
-            using (var host = CreateOmniSharpHost(testFile))
+            var request = new FormatRangeRequest()
             {
-                var span = testFile.Content.GetSpans().Single();
-                var range = testFile.Content.GetRangeFromSpan(span);
+                Buffer = testFile.Content.Code,
+                FileName = testFile.FileName,
+                Line = range.Start.Line,
+                Column = range.Start.Offset,
+                EndLine = range.End.Line,
+                EndColumn = range.End.Offset
+            };
 
-                var request = new FormatRangeRequest()
-                {
-                    Buffer = testFile.Content.Code,
-                    FileName = testFile.FileName,
-                    Line = range.Start.Line,
-                    Column = range.Start.Offset,
-                    EndLine = range.End.Line,
-                    EndColumn = range.End.Offset
-                };
+            var requestHandler = SharedOmniSharpTestHost.GetRequestHandler<FormatRangeService>(OmniSharpEndpoints.FormatRange);
 
-                var requestHandler = host.GetRequestHandler<FormatRangeService>(OmniSharpEndpoints.FormatRange);
+            var response = await requestHandler.Handle(request);
+            var actual = response.Changes.ToArray();
 
-                var response = await requestHandler.Handle(request);
-                var actual = response.Changes.ToArray();
+            Assert.Equal(expected.Length, actual.Length);
 
-                Assert.Equal(expected.Length, actual.Length);
-
-                for (var i = 0; i < expected.Length; i++)
-                {
-                    Assert.Equal(expected[i].NewText, actual[i].NewText);
-                    Assert.Equal(expected[i].StartLine, actual[i].StartLine);
-                    Assert.Equal(expected[i].StartColumn, actual[i].StartColumn);
-                    Assert.Equal(expected[i].EndLine, actual[i].EndLine);
-                    Assert.Equal(expected[i].EndColumn, actual[i].EndColumn);
-                }
+            for (var i = 0; i < expected.Length; i++)
+            {
+                Assert.Equal(expected[i].NewText, actual[i].NewText);
+                Assert.Equal(expected[i].StartLine, actual[i].StartLine);
+                Assert.Equal(expected[i].StartColumn, actual[i].StartColumn);
+                Assert.Equal(expected[i].EndLine, actual[i].EndLine);
+                Assert.Equal(expected[i].EndColumn, actual[i].EndColumn);
             }
         }
 
@@ -306,31 +303,29 @@ class C {
         {
             var testFile = new TestFile("dummy.cs", source);
 
-            using (var host = CreateOmniSharpHost(testFile))
+            SharedOmniSharpTestHost.AddFilesToWorkspace(testFile);
+            var span = testFile.Content.GetSpans().Single();
+            var range = testFile.Content.GetRangeFromSpan(span);
+
+            var request = new FormatRangeRequest()
             {
-                var span = testFile.Content.GetSpans().Single();
-                var range = testFile.Content.GetRangeFromSpan(span);
+                Buffer = testFile.Content.Code,
+                FileName = testFile.FileName,
+                Line = range.Start.Line,
+                Column = range.Start.Offset,
+                EndLine = range.End.Line,
+                EndColumn = range.End.Offset
+            };
 
-                var request = new FormatRangeRequest()
-                {
-                    Buffer = testFile.Content.Code,
-                    FileName = testFile.FileName,
-                    Line = range.Start.Line,
-                    Column = range.Start.Offset,
-                    EndLine = range.End.Line,
-                    EndColumn = range.End.Offset
-                };
+            var requestHandler = SharedOmniSharpTestHost.GetRequestHandler<FormatRangeService>(OmniSharpEndpoints.FormatRange);
 
-                var requestHandler = host.GetRequestHandler<FormatRangeService>(OmniSharpEndpoints.FormatRange);
+            var response = await requestHandler.Handle(request);
+            var actual = response.Changes.ToArray();
 
-                var response = await requestHandler.Handle(request);
-                var actual = response.Changes.ToArray();
-
-                var oldText = testFile.Content.Text;
-                var textChanges = GetTextChanges(oldText, response.Changes);
-                var actualText = oldText.WithChanges(textChanges).ToString();
-                Assert.Equal(expected.Replace("\r\n","\n"), actualText.Replace("\r\n","\n"));
-            }
+            var oldText = testFile.Content.Text;
+            var textChanges = GetTextChanges(oldText, response.Changes);
+            var actualText = oldText.WithChanges(textChanges).ToString();
+            Assert.Equal(expected.Replace("\r\n", "\n"), actualText.Replace("\r\n", "\n"));
         }
 
         public static IEnumerable<TextChange> GetTextChanges(SourceText oldText, IEnumerable<LinePositionSpanTextChange> changes)